moved game2 components into activity1
