+++ src/components/Activities/Activity1/Game2/components/GameSlide.tsx
@@ -1,13 +1,16 @@
 import React, { useState } from 'react';
 
 // import Anime from 'react-anime';
-// import useSound from 'use-sound';
+import useSound from 'use-sound';
 
-import AlienSvg from '../../../../assets/activity1/game2/alien.svg';
-// import CorrectSFX from '../../../../assets/activity1/game2/anime-wow-sound-effect.mp3';
-// import IncorrectSFX from '../../../../assets/activity1/game2/oh_no_1.mp3';
+import AlienSvg from '../../../../../assets/activity1/game2/alien.svg';
+import CorrectSFX from '../../../../../assets/activity1/game2/correct.mp3';
+import IncorrectSFX from '../../../../../assets/activity1/game2/oh_no_1.mp3';
 
+import { TextBubbleStyles } from '../../../../shared/PlaynetConstants';
+import TextBubble from '../../TextBubble';
 import SlideBox from './SlideBox';
+
 
 interface GameSlideProps {
   correctImg ?: number;
@@ -20,28 +23,26 @@ interface GameSlideProps {
 
 function GameSlide(props: GameSlideProps): JSX.Element {
   const [incorrect, setIncorrect] = useState(false);
-  // const [playCorrect] = useSound(CorrectSFX, { volume: 0.5});
-  // const [playIncorrect] = useSound(IncorrectSFX, { volume: 0.5});
-
-  const defaultError = 'error: missing text';
+  const [playCorrect] = useSound(CorrectSFX, { volume: 0.01});
+  const [playIncorrect] = useSound(IncorrectSFX, { volume: 0.01});
 
   const img0 = props.img0;
   const img1 = props.img1;
 
   const displayText: () => string = () => {
-    if (incorrect) return props.textIncorrect ? props.textIncorrect : defaultError;
-    return props.textDefault ? props.textDefault : defaultError;
+    if (incorrect) return props.textIncorrect ? props.textIncorrect : '';
+    return props.textDefault ? props.textDefault : '';
   };
 
   const handleClick = (pos : number) => {
     let newIncorrect = true;
 
     if (pos === props.correctImg) {
-      // playCorrect();
+      playCorrect();
       newIncorrect = false;
       props.advanceGame && props.advanceGame();
     } else if (!incorrect) {
-      // playIncorrect();
+      playIncorrect();
     }
     setIncorrect(newIncorrect);
   };
@@ -49,7 +50,7 @@ function GameSlide(props: GameSlideProps): JSX.Element {
   return (
     <div id={'game-content'}>
       <div className={'gamebox'}>
-        <h3> {displayText()} </h3>
+        <TextBubble textBubbleStyle={TextBubbleStyles.EXTRA_LARGE} text={displayText()} />
         <img src={AlienSvg} />
       </div>
       <div className={'gamebox'}>
