+++ src/components/Activities/Activity1/Game2/components/SlideBox.tsx
