+++ src/components/styles/Activity1.scss
@@ -74,3 +74,50 @@
   left: 50%;
   opacity: 0;
 }
+
+#game2-intro {
+  display: flex;
+  height: 100%;
+  flex-flow: column wrap;
+  justify-content: center;
+  align-items: center;
+}
+
+#game2-intro-button {
+  width: 300px;
+  height: 50px;
+  max-width: 50%;
+  max-height: 50%;
+  background-color: #ff5c41;
+  color: white;
+  font-size: 1vw;
+  font-family: inherit;
+  border-radius: 5px;
+  border: none;
+}
+
+#blank-computer {
+  display: inline-block;
+  position: relative;
+  width: 50%;
+}
+
+#blank-computer-img {
+  width: 100%;
+}
+
+#blank-computer-text {
+  position: absolute;
+  top: 40%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+  font-size: 1.3vw;
+  text-align: left;
+  width: 80%;
+  max-height: 70%;
+}
+
+#blank-computer-title-text {
+  text-align: center;
+  color: red;
+}
